+++ lib/src/utils/iterable_extensions.dart
@@ -1,5 +1,25 @@
 library utils.list_extensions;
 
+import 'package:build_html/build_html.dart';
+
+extension HtmlExtension on List<List<String>> {
+  Table toHtml() {
+    var table = Table();
+    var header = TableRow();
+    for (var colNames in first) {
+      header.addCell(TableCell(TableCellType.header)..withRaw(colNames));
+    }
+    for (var cells in skip(1)) {
+      var row = TableRow();
+      for (var cell in cells) {
+        row.addCell(TableCell(TableCellType.data)..withRaw(cell));
+      }
+      table.addCustomBodyRow(row);
+    }
+    return table;
+  }
+}
+
 extension IterableExtensions<E> on Iterable<E> {
   /// Partition an iterable based on a condition.
   /// First element (left) is the true condition, the second element (right)
@@ -16,5 +36,4 @@ extension IterableExtensions<E> on Iterable<E> {
     }
     return (left, right);
   }
-
 }